mmbc: no %lld - the Fuzix libc silently eats it
The host build printed 64-bit values with %lld and every gate passed;
the ARM build's printf has no %lld, consumed the argument list
wrongly, and the board's mmbc emitted mm_ssetm calls with the string
argument MISSING - caught on hardware when type.bas would not
compile.  Every 64-bit format in the TYPE code now casts to (long)
with %ld; the values are struct offsets, sizes and dims, all far
inside 32 bits, so the emitted text is unchanged (cgate 0).

fcc/genexpect.sh regenerates the two struct .expected files from
host-armm0 bcrun runs.

Board-verified self-hosted: type.bas prints the same fifteen lines as
the host - the layout constants 16/56/31/120 included - and the
firmware's StructTest.bas translates with the identical 81-line
refusal report, compiles, and scores the same 58 PASS / 45 FAIL.

+++ mmbc/mmbc_out.c
@@ -361,26 +361,26 @@ void conv_write(FILE *f)
                 struct typemember *m = td->members[j];
                 const char *decl;
                 if (m->offset > pos) {
-                    fprintf(f, "    unsigned char __p%d[%lld];\n",
-                            padn, m->offset - pos);
+                    fprintf(f, "    unsigned char __p%d[%ld];\n",
+                            padn, (long)(m->offset - pos));
                     padn++;
                 }
                 if (m->stype != NULL) {
                     decl = sfmt("struct t_%s m_%s", m->stype, m->name);
                     if (m->has_dims)
-                        decl = sfmt("%s[%lld]", decl, m->count);
+                        decl = sfmt("%s[%ld]", decl, (long)m->count);
                 } else if (m->ty == TY_S) {
-                    decl = sfmt("char m_%s[%lld]", m->name,
-                                m->esize * m->count);
+                    decl = sfmt("char m_%s[%ld]", m->name,
+                                (long)(m->esize * m->count));
                 } else {
                     decl = sfmt("%s m_%s", ctype_of(m->ty), m->name);
                     if (m->has_dims)
-                        decl = sfmt("%s[%lld]", decl, m->count);
+                        decl = sfmt("%s[%ld]", decl, (long)m->count);
                 }
                 fprintf(f, "    %s;\n", decl);
                 pos = m->offset + m->esize * m->count;
             }
-            fprintf(f, "};    /* %lld bytes */\n", td->total);
+            fprintf(f, "};    /* %ld bytes */\n", (long)td->total);
         }
         fprintf(f, "\n");
     }
